App now successfully merged, but requires the index IX_Tags_UserId to be set to non-unique AFTER the database has been created
